refactor(DEV-11087): remove unnecessary reexport
