--- packages/sdk-react/src/components/approvalRequests/ApprovalRequestsTable/ApprovalRequestsFilter/index.ts
@@ -1,1 +1,0 @@
-export { ApprovalRequestsFilter } from './ApprovalRequestsFilter';
