--- packages/sdk-react/src/components/approvalRequests/ApprovalRequestsTable/ApproveButton/index.ts
@@ -1,1 +1,0 @@
-export { ApproveButton } from './ApproveButton';
