--- packages/sdk-react/src/components/approvalRequests/ApprovalRequestsTable/AutocompleteCreatedBy/index.ts
@@ -1,1 +1,0 @@
-export { AutocompleteCreatedBy } from './AutocompleteCreatedBy';
