--- packages/sdk-react/src/components/approvalRequests/ApprovalRequestsTable/RejectButton/index.ts
@@ -1,1 +1,0 @@
-export { RejectButton } from './RejectButton';
