--- packages/sdk-react/src/components/approvalRequests/ApprovalRequestsTable/UserCell/index.ts
@@ -1,1 +1,0 @@
-export { UserCell } from './UserCell';
